feat: add breakPoints const to constants folder
