+++ src/constants/sliderBreakPoints.js
@@ -1,0 +1,8 @@
+const breakPoints = [
+  { width: 1, itemsToShow: 1, pagination: false },
+  { width: 400, itemsToShow: 2 },
+  { width: 750, itemsToShow: 3 },
+  { width: 900, itemsToShow: 4 },
+];
+
+export default breakPoints;
